Rewrite Step1 to ts
