+++ src/app/components/shared/radio-list/radioList.less
@@ -1,0 +1,40 @@
+@import '../../../../../node_modules/nav-frontend-core/less/_variabler';
+
+.radioList {
+  &__radio {
+    margin-top: 0.7rem;
+    margin-bottom: 0.5rem;
+    label {
+      font-weight: normal;
+    }
+  }
+
+  .skjemaelement {
+    &__label {
+      width: 100%;
+      padding: 1rem 1rem 1rem 50px;
+      line-height: 24px;
+      background-color: @white;
+      border: 1px solid @navGra20;
+      border-radius: 4px;
+      transition: border-color 200ms cubic-bezier(0.465, 0.183, 0.153, 0.946);
+      &:hover {
+        border-color: #0067c5;
+      }
+    }
+
+    &__label:before {
+      border-radius: 50%;
+      width: 24px;
+      height: 24px;
+      top: 50%;
+      margin-top: -12px;
+      margin-left: 12px;
+    }
+
+    &__label:after {
+      border-radius: 50%;
+      margin-right: 12px;
+    }
+  }
+}
